services/integration-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
pika
python-dotenv
